Class

+++ Class/Class.Binary/Read.cs
@@ -596,13 +596,13 @@ public class Read : Any
 
     protected virtual Array ExecutePartArray()
     {
-        int o;
+        long o;
         o = this.ExecuteCount();
         if (o == -1)
         {
             return null;
         }
-        int count;
+        long count;
         count = o;
 
         Array array;
@@ -612,7 +612,7 @@ public class Read : Any
             return null;
         }
 
-        int i;
+        long i;
         i = 0;
         while (i < count)
         {
